Sunset read-only serializer
